MainActivity очищен от мусора путём переноса UI MainScreen Убрано много дублирующегося кода с помощью абстрацкий BaseLoadViewModel и Composable LoadList
