+++ app/src/main/java/com/gena_korobeynikov/yandexfinance/data/repo_Implementations/AccountRepositoryImpl.kt
@@ -1,8 +1,9 @@
-package com.gena_korobeynikov.yandexfinance.domain
+package com.gena_korobeynikov.yandexfinance.data.repo_Implementations
 
 import com.gena_korobeynikov.yandexfinance.data.api.AccountApi
-import com.gena_korobeynikov.yandexfinance.data.api.TransactionsApi
 import com.gena_korobeynikov.yandexfinance.data.toDomain
+import com.gena_korobeynikov.yandexfinance.domain.Account
+import com.gena_korobeynikov.yandexfinance.domain.AccountRepository
 
 class AccountRepositoryImpl (
     private val api: AccountApi
