+++ app/src/main/java/com/gena_korobeynikov/yandexfinance/ui/UiState.kt
@@ -1,0 +1,7 @@
+package com.gena_korobeynikov.yandexfinance.ui
+
+sealed class UiState<out T> {
+    data object Loading : UiState<Nothing>()
+    data class Success<T>(val data: T) : UiState<T>()
+    data class Error(val message: String) : UiState<Nothing>()
+}
