+++ app/src/main/java/com/gena_korobeynikov/yandexfinance/ui/screens/AccountScreen.kt
@@ -2,14 +2,11 @@ package com.gena_korobeynikov.yandexfinance.ui.screens
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
@@ -20,23 +17,19 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.gena_korobeynikov.yandexfinance.R
 import com.gena_korobeynikov.yandexfinance.common.NetworkModule
 import com.gena_korobeynikov.yandexfinance.data.toSymbol
 import com.gena_korobeynikov.yandexfinance.domain.Account
-import com.gena_korobeynikov.yandexfinance.domain.AccountRepositoryImpl
-import com.gena_korobeynikov.yandexfinance.domain.TransactionsRepositoryImpl
+import com.gena_korobeynikov.yandexfinance.data.repo_Implementations.AccountRepositoryImpl
+import com.gena_korobeynikov.yandexfinance.ui.UiState
+import com.gena_korobeynikov.yandexfinance.ui.components.ListLoader
 import com.gena_korobeynikov.yandexfinance.ui.components.MainListItem
 import com.gena_korobeynikov.yandexfinance.ui.states.AccountUiState
-import com.gena_korobeynikov.yandexfinance.ui.states.TransactionUiState
 import com.gena_korobeynikov.yandexfinance.ui.viewModels.AccountViewModel
-import com.gena_korobeynikov.yandexfinance.ui.viewModels.TransactionsViewModel
 
 
 @Composable
@@ -54,32 +47,12 @@ import com.gena_korobeynikov.yandexfinance.ui.viewModels.TransactionsViewModel
         viewModel.loadAccount(accountId)
     }
 
-    when (uiState) {
-        is AccountUiState.Loading -> {
-            Box(
-                modifier = Modifier.fillMaxSize(),
-                contentAlignment = Alignment.Center
-            ) {
-                CircularProgressIndicator()
-            }
-        }
 
-        is AccountUiState.Success -> {
-            val account = (uiState as AccountUiState.Success).account
-            AccountInfo(account)
-        }
-
-        is AccountUiState.Error -> {
-            val message = (uiState as AccountUiState.Error).message
-            Box(
-                modifier = Modifier.fillMaxSize(),
-                contentAlignment = Alignment.Center
-            ) {
-                Text(text = "Ошибка: $message", color = Color.Red)
-            }
-        }
-
+    ListLoader(uiState) {
+        val account = (uiState as UiState.Success).data
+        AccountInfo(account)
     }
+
 }
