+++ app/src/main/java/com/gena_korobeynikov/yandexfinance/ui/screens/CategoriesScreen.kt
@@ -2,14 +2,12 @@
 
 package com.gena_korobeynikov.yandexfinance.ui.screens
 
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -23,7 +21,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.RectangleShape
@@ -33,11 +30,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.gena_korobeynikov.yandexfinance.R
 import com.gena_korobeynikov.yandexfinance.common.NetworkModule
-import com.gena_korobeynikov.yandexfinance.domain.CategoriesRepositoryImpl
+import com.gena_korobeynikov.yandexfinance.data.repo_Implementations.CategoriesRepositoryImpl
 import com.gena_korobeynikov.yandexfinance.domain.Category
+import com.gena_korobeynikov.yandexfinance.ui.UiState
+import com.gena_korobeynikov.yandexfinance.ui.components.ListLoader
 import com.gena_korobeynikov.yandexfinance.ui.components.MainListItem
 import com.gena_korobeynikov.yandexfinance.ui.states.CategoryUiState
-import com.gena_korobeynikov.yandexfinance.ui.states.TransactionUiState
 import com.gena_korobeynikov.yandexfinance.ui.viewModels.CategoriesViewModel
 
 
@@ -56,35 +54,17 @@ import com.gena_korobeynikov.yandexfinance.ui.viewModels.CategoriesViewModel
         viewModel.loadCategories()
     }
 
-    when (uiState) {
-        is CategoryUiState.Loading -> {
-            Box(
-                modifier = Modifier.fillMaxSize(),
-                contentAlignment = Alignment.Center
-            ) {
-                CircularProgressIndicator()
-            }
-        }
 
-        is CategoryUiState.Success -> {
-            val categories = (uiState as CategoryUiState.Success).categories
-            Column {
-                SearchField()
-                CategoryList(categories)
-            }
+    ListLoader(
+        uiState
+    ) {
+        val categories = (uiState as UiState.Success).data
+        Column {
+            SearchField()
+            CategoryList(categories)
         }
-
-        is CategoryUiState.Error -> {
-            val message = (uiState as CategoryUiState.Error).message
-            Box(
-                modifier = Modifier.fillMaxSize(),
-                contentAlignment = Alignment.Center
-            ) {
-                Text(text = "Ошибка: $message", color = Color.Red)
-            }
-        }
-
     }
+
 }
