+++ app/src/main/java/com/gena_korobeynikov/yandexfinance/ui/screens/HistoryScreen.kt
@@ -3,13 +3,11 @@ package com.gena_korobeynikov.yandexfinance.ui.screens
 import android.app.DatePickerDialog
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
@@ -22,7 +20,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
@@ -32,7 +29,9 @@ import com.gena_korobeynikov.yandexfinance.R
 import com.gena_korobeynikov.yandexfinance.common.NetworkModule
 import com.gena_korobeynikov.yandexfinance.data.toSymbol
 import com.gena_korobeynikov.yandexfinance.domain.Transaction
-import com.gena_korobeynikov.yandexfinance.domain.TransactionsRepositoryImpl
+import com.gena_korobeynikov.yandexfinance.data.repo_Implementations.TransactionsRepositoryImpl
+import com.gena_korobeynikov.yandexfinance.ui.UiState
+import com.gena_korobeynikov.yandexfinance.ui.components.ListLoader
 import com.gena_korobeynikov.yandexfinance.ui.components.MainListItem
 import com.gena_korobeynikov.yandexfinance.ui.states.TransactionUiState
 import com.gena_korobeynikov.yandexfinance.ui.viewModels.TransactionsViewModel
@@ -135,49 +134,26 @@ fun HistoryScreen(isIncomes : Boolean = false) {
         )
 
 
+        ListLoader(uiState) {
+            val expenses =
+                (uiState as UiState.Success).data.filter { it.category.isIncome == isIncomes }
+                    .sortedByDescending { it.transactionDate }
 
-
-        when (uiState) {
-            is TransactionUiState.Loading -> {
-                Box(
-                    modifier = Modifier.fillMaxSize(),
-                    contentAlignment = Alignment.Center
-                ) {
-                    CircularProgressIndicator()
+            MainListItem(
+                mainText = stringResource(id = R.string.sum),
+                color = colorResource(id = R.color.secondary_green),
+                huggingHeight = true,
+                trailing = {
+                    Text(
+                        text = "${expenses.sumOf { it.amount }} ${expenses.firstOrNull()?.account?.currency?.toSymbol() ?: ""}",
+                        style = MaterialTheme.typography.bodyLarge,
+                    )
                 }
-            }
-
-            is TransactionUiState.Success -> {
-                val expenses =
-                    (uiState as TransactionUiState.Success).transactions.filter { it.category.isIncome == isIncomes }
-                        .sortedByDescending { it.transactionDate }
-
-
-                MainListItem(
-                    mainText = stringResource(id = R.string.sum),
-                    color = colorResource(id = R.color.secondary_green),
-                    huggingHeight = true,
-                    trailing = {
-                        Text(
-                            text = "${expenses.sumOf { it.amount }} ${expenses.firstOrNull()?.account?.currency?.toSymbol()?: ""}",
-                            style = MaterialTheme.typography.bodyLarge,
-                        )
-                    }
-                )
-                HistoryList(expenses)
-            }
-
-            is TransactionUiState.Error -> {
-                val message = (uiState as TransactionUiState.Error).message
-                Box(
-                    modifier = Modifier.fillMaxSize(),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Text(text = "Ошибка: $message", color = Color.Red)
-                }
-            }
-
+            )
+            HistoryList(expenses)
         }
+
+
     }
 }
