+++ app/src/main/java/com/gena_korobeynikov/yandexfinance/ui/screens/SettingsScreen.kt
@@ -1,30 +1,19 @@
 package com.gena_korobeynikov.yandexfinance.ui.screens
 
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.KeyboardArrowRight
-import androidx.compose.material.icons.filled.PlayArrow
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material.icons.materialIcon
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Switch
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.util.fastCbrt
 import com.gena_korobeynikov.yandexfinance.R
 import com.gena_korobeynikov.yandexfinance.ui.components.MainListItem
 
 
-    @Composable
+@Composable
     fun SettingsScreen() {
         val itemTitles = listOf(
             stringResource(id = R.string.main_color),
